# internal/models/todo.go
package models

type Todo struct {
    ID        int    `json:"id"`
    Title     string `json:"title"`
    Completed bool   `json:"completed"`
    CreatedAt string `json:"created_at"`
}

type CreateTodoInput struct {
    Title string `json:"title"`
}

type UpdateTodoInput struct {
    Title     *string `json:"title"`
    Completed *bool   `json:"completed"`
}